Sync processing rewritten and some issues fixed
* add 480i/576i line3x/4x
* clean up RTL and SDC
* update AVI infoframe properly
* add HDMI IT content option

+++ software/sys_controller/it6613/it6613_drv.h
@@ -834,7 +834,7 @@ BOOL EnableAudioInfoFrame(BYTE bEnable,BYTE *pAudioInfoFrame);
 void SetAVMute(BYTE bEnable) ;
 void SetOutputColorDepthPhase(BYTE ColorDepth,BYTE bPhase) ;
 void Get6613Reg(BYTE *pReg) ;
-void HDMITX_SetPixelRepetition(int pixelrep, int set_infoframe);
+void HDMITX_SetPixelRepetition(BYTE pixelrep, BYTE set_infoframe);
 
 ////////////////////////////////////////////////////////////////////
 // Required Interfance
